sokol_gfx.h: fix msvc errors

+++ sokol_gfx.h
@@ -17147,7 +17147,7 @@ SOKOL_API_IMPL void sg_apply_uniforms(sg_shader_stage stage, int ub_index, const
     SOKOL_ASSERT((ub_index >= 0) && (ub_index < SG_MAX_SHADERSTAGE_UBS));
     SOKOL_ASSERT(data && data->ptr && (data->size > 0));
     _sg.stats.num_apply_uniforms++;
-    _sg.stats.size_apply_uniforms += data->size;
+    _sg.stats.size_apply_uniforms += (uint32_t)data->size;
     if (!_sg_validate_apply_uniforms(stage, ub_index, data)) {
         _sg.next_draw_valid = false;
         return;
@@ -17226,7 +17226,7 @@ SOKOL_API_IMPL void sg_update_buffer(sg_buffer buf_id, const sg_range* data) {
     SOKOL_ASSERT(_sg.valid);
     SOKOL_ASSERT(data && data->ptr && (data->size > 0));
     _sg.stats.num_update_buffer++;
-    _sg.stats.size_update_buffer += data->size;
+    _sg.stats.size_update_buffer += (uint32_t)data->size;
     _sg_buffer_t* buf = _sg_lookup_buffer(&_sg.pools, buf_id.id);
     if ((data->size > 0) && buf && (buf->slot.state == SG_RESOURCESTATE_VALID)) {
         if (_sg_validate_update_buffer(buf, data)) {
@@ -17246,7 +17246,7 @@ SOKOL_API_IMPL int sg_append_buffer(sg_buffer buf_id, const sg_range* data) {
     SOKOL_ASSERT(_sg.valid);
     SOKOL_ASSERT(data && data->ptr);
     _sg.stats.num_append_buffer++;
-    _sg.stats.size_append_buffer += data->size;
+    _sg.stats.size_append_buffer += (uint32_t)data->size;
     _sg_buffer_t* buf = _sg_lookup_buffer(&_sg.pools, buf_id.id);
     int result;
     if (buf) {
@@ -17314,7 +17314,7 @@ SOKOL_API_IMPL void sg_update_image(sg_image img_id, const sg_image_data* data) 
             if (data->subimage[face_index][mip_index].size == 0) {
                 break;
             }
-            _sg.stats.size_update_image += data->subimage[face_index][mip_index].size;
+            _sg.stats.size_update_image += (uint32_t)data->subimage[face_index][mip_index].size;
         }
     }
     _sg_image_t* img = _sg_lookup_image(&_sg.pools, img_id.id);
